requirements.txt
streamlit==1.34.0
scikit-learn==1.4.2
xgboost==1.5.2
PyYAML==6.0.1
pandas==2.2.2
numpy==1.26.4
pillow==10.3.0



